requirements.txt
accelerate==1.8.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
attrs==25.3.0
beautifulsoup4==4.13.4
bidict==0.23.1
blinker==1.9.0
blis==1.3.0
captum==0.8.0
catalogue==2.0.10
certifi==2025.6.15
charset-normalizer==3.4.2
choreographer==1.0.9
click==8.2.1
cloudpathlib==0.21.1
colorama==0.4.6
confection==0.1.5
contourpy==1.3.2
cycler==0.12.1
cymem==2.0.11
datasets==3.6.0
decorator==5.2.1
deep-translator==1.11.4
diffusers==0.34.0
dill==0.3.8
distro==1.9.0
docstring_parser==0.16
dotenv==0.9.9
einops==0.8.1
entmax==1.3
et_xmlfile==2.0.0
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.18.0
Flask==3.1.1
flatten-dict==0.4.2
fonttools==4.58.4
frozenlist==1.7.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.2
idna==3.10
importlib_metadata==8.7.0
inseq==0.6.0
ipython==9.3.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jaxtyping==0.3.2
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonargparse==3.13.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_core==5.8.1
kaleido==1.0.0
kiwisolver==1.4.8
langcodes==3.5.0
langdetect==1.0.9
language_data==1.3.0
lightning-utilities==0.14.3
logistro==1.1.0
lxml==5.4.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
msgspec==0.19.0
multidict==6.6.3
multiprocess==0.70.16
murmurhash==1.0.13
narwhals==1.45.0
nbformat==5.10.4
networkx==3.5
nltk==3.9.1
nnsight==0.4.8
numpy==2.3.1
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
openai==1.93.0
openpyxl==3.1.5
orjson==3.10.18
packaging==25.0
pandas==2.3.0
parso==0.8.4
patsy==1.0.1
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
plotly==6.2.0
portalocker==3.2.0
preshed==3.0.10
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-engineio==4.12.2
python-socketio==5.13.0
pytorch-lightning==2.5.1.post0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
rpds-py==0.25.1
sacrebleu==2.5.1
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
seaborn==0.13.2
sentence-transformers==4.1.0
sentencepiece==0.2.0
sentry-sdk==2.32.0
setuptools==80.9.0
shellingham==1.5.4
simple-parsing==0.1.7
simple-websocket==1.1.0
simplejson==3.20.1
six==1.17.0
smart_open==7.3.0.post1
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
stack-data==0.6.3
statsmodels==0.14.5
sympy==1.14.0
tabulate==0.9.0
thinc==8.3.6
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.2
toml==0.10.2
torch==2.7.1
torchdata==0.11.0
torchmetrics==0.10.3
torchvision==0.22.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.53.0
triton==3.3.1
tuned-lens==0.2.0
typeguard==2.13.3
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
unbabel-comet==2.2.6
UpSetPlot==0.9.0
urllib3==2.5.0
uv==0.7.13
wadler_lindig==0.1.7
wandb==0.21.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
wsproto==1.2.0
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
